% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{boxPlot}
\alias{boxPlot}
\title{Plot box and whisker of marks and grade boundaries}
\usage{
boxPlot(data, x, y, yLabel, boundaries, cutScores, gradeLabels)
}
\arguments{
\item{data}{results data frame}

\item{x}{x co-ordinate column name}

\item{y}{y co-ordinate column name}

\item{boundaries}{grade boundaries data frame}

\item{cutScores}{cut scores for horizontal intersect}

\item{gradeLabels}{labels for cut scores}

\item{yLable}{label for y axis}
}
\description{
Plot box and whisker of marks and grade boundaries
}
\examples{
p <- boxPlot(results, x=school, y=scaledScore, yLabel=mark, boundaries=displayGrades, cutScores=scaledScore, gradeLabels=gradeBoundaries)
}
