% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{cleanJudges}
\alias{cleanJudges}
\title{Set judges to exclude on national moderation tasks}
\usage{
cleanJudges(taskName, connStr, infit = 1.3, decisions = 20)
}
\arguments{
\item{taskName}{name of tasks (uses regex)}

\item{connStr}{mongo connection string}
}
\value{
nothing
}
\description{
Set judges to exclude on national moderation tasks
}
\examples{
cleanJudges(taskName, connStr, infit, decisions)
}
