% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{error.bar}
\alias{error.bar}
\title{Plot pupil score with error and grade boundaries}
\usage{
error.bar(pupil, grade.boundaries, errorScale)
}
\arguments{
\item{pupil}{results row from results data frame}

\item{grade.boundaries}{grade boundary data frame}

\item{errorScale}{vector for x scale limits}
}
\value{
plot
}
\description{
Plot pupil score with error and grade boundaries
}
\examples{
p <- error.bar(pupil, grade.boundaries, errorScale)
}
