% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{genderPlot}
\alias{genderPlot}
\title{Plot data with mean and confidence intervals, grouped by gender}
\usage{
genderPlot(results, ylimits, ylabel)
}
\arguments{
\item{results}{results data frame}

\item{ylimits}{limits for y axis}

\item{ylabel}{label for y axis}
}
\value{
plot
}
\description{
Plot data with mean and confidence intervals, grouped by gender
}
\examples{
summary.data <- genderPlot(school.results)
}
