% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{getData}
\alias{getData}
\title{Get data from a series of national tasks, and clean}
\usage{
getData(taskName, connStr)
}
\arguments{
\item{taskName}{name of tasks (uses regex)}

\item{connStr}{mongo connection string}
}
\value{
data frame results
}
\description{
Get data from a series of national tasks, and clean
}
\examples{
results <- getData(taskName, connStr)
}
