% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{meanSePlot}
\alias{meanSePlot}
\title{Plot data with mean and confidence intervals}
\usage{
meanSePlot(data, x, ylimits, ylabel, data.mean = data.mean,
  data.lower = data.lower, data.upper = data.upper, fill = NULL)
}
\arguments{
\item{data}{results data frame}

\item{x}{column for x co-ordinate}

\item{ylimits}{limits for y axis}

\item{ylabel}{label for y axis}

\item{data.mean}{column for y co-ordinate}

\item{data.lower}{column for lower error bar}

\item{data.upper}{column for upper error bar}

\item{fill}{grouping variable for fill}
}
\value{
plot
}
\description{
Plot data with mean and confidence intervals
}
\examples{
p <- meanSePlot(pp.summary, ylimits=ylimits, ylabel=ylabel,x = school, fill=pp)
}
