% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{ppPlot}
\alias{ppPlot}
\title{Plot data with mean and confidence intervals, grouped by pupil premium}
\usage{
ppPlot(results, ylimits, ylabel)
}
\arguments{
\item{results}{results data frame}

\item{ylimits}{limits for y axis}

\item{ylabel}{label for y axis}
}
\value{
plot
}
\description{
Plot data with mean and confidence intervals, grouped by pupil premium
}
\examples{
summary.data <- ppPlot(school.results)
}
