% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{schools.gender.plot}
\alias{schools.gender.plot}
\title{Headline box and whiskers school compared to cohort by gender}
\usage{
schools.gender.plot(results, grade.boundaries)
}
\arguments{
\item{results}{results data frame}

\item{grade.boundaries}{grade boundaries}
}
\value{
plot
}
\description{
Headline box and whiskers school compared to cohort by gender
}
\examples{
p <- schools.gender.plot(results, grade.boundaries)
}
