% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{schools.plot}
\alias{schools.plot}
\title{Headline box and whiskers school compared to all other schools individually}
\usage{
schools.plot(results, yname, grade.boundaries = NULL)
}
\arguments{
\item{results}{results data frame}

\item{yname}{label for y axis}

\item{grade.boundaries}{grade boundaries}
}
\value{
plot
}
\description{
Headline box and whiskers school compared to all other schools individually
}
\examples{
p <- schools.plot(results, grade.boundaries)
}
