% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{scoreSignificanceStatement}
\alias{scoreSignificanceStatement}
\title{Create significance statement for reports using scaled scores}
\usage{
scoreSignificanceStatement(results)
}
\arguments{
\item{results}{results data frame}
}
\value{
significance statement
}
\description{
Create significance statement for reports using scaled scores
}
\examples{
all.significance.statement <- scoreSignificanceStatement(results)
}
