% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{significanceStatement}
\alias{significanceStatement}
\title{Create significance statement for reports using grades}
\usage{
significanceStatement(results)
}
\arguments{
\item{results}{results data frame}
}
\value{
significance statement
}
\description{
Create significance statement for reports using grades
}
\examples{
all.significance.statement <- significanceStatement(results)
}
