% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{summarisePPData}
\alias{summarisePPData}
\title{Summarise data with mean and confidence intervals, grouped on pupil premium}
\usage{
summarisePPData(results)
}
\arguments{
\item{results}{results data frame}
}
\value{
data frame with n, mean, sd, se, ci, upper and lower
}
\description{
Summarise data with mean and confidence intervals, grouped on pupil premium
}
\examples{
summary.data <- summarisePPData(school.results)
}
