% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{summaryPlot}
\alias{summaryPlot}
\title{Plot data with mean and confidence intervals}
\usage{
summaryPlot(results, ylimits, ylabel)
}
\arguments{
\item{results}{results data frame}

\item{ylimits}{limits for y axis}

\item{ylabel}{label for y axis}
}
\value{
plot
}
\description{
Plot data with mean and confidence intervals
}
\examples{
summary.data <- summaryPlot(school.results)
}
