% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{summariseData}
\alias{summariseData}
\title{Summarise data with mean and confidence intervals}
\usage{
summariseData(data, by, var)
}
\arguments{
\item{data}{data frame}

\item{by}{the name of the grouping variable}

\item{var}{the name of the variable to summarise}
}
\value{
data frame with n, mean, sd, se, ci, upper and lower
}
\description{
Summarise data with mean and confidence intervals
}
\examples{
summary.data <- summaryData(school.results, school, gradeScore)
}
