% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{summariseData}
\alias{summariseData}
\title{Summarise data with mean and confidence intervals, grouped on school}
\usage{
summariseData(results)
}
\arguments{
\item{results}{results data frame}
}
\value{
data frame with n, mean, sd, se, ci, upper and lower
}
\description{
Summarise data with mean and confidence intervals, grouped on school
}
\examples{
summary.data <- summaryData(school.results)
}
